pyproject.toml
[project]
name = "chart"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-genai>=1.62.0",
    "mplfinance>=0.12.10b0",
    "pandas>=3.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pandas-stubs>=2.2.0",
    "types-requests>=2.32.0",
]

[tool.mypy]
python_version = "3.14"
warn_return_any = true
warn_unused_configs = true
check_untyped_defs = true
ignore_missing_imports = true
